feat(clock_angles_dto): Выделение ClockAngles в отдельный файл
